+++ src/mako/http/routing/traits/ControllerHelperTrait.php
@@ -8,9 +8,9 @@
 namespace mako\http\routing\traits;
 
 use Closure;
+use Deprecated;
 use mako\http\response\builders\JSON;
 use mako\http\response\senders\File;
-use mako\http\response\senders\Redirect;
 use mako\http\response\senders\Stream;
 use mako\syringe\traits\ContainerAwareTrait;
 
@@ -20,30 +20,23 @@ use mako\syringe\traits\ContainerAwareTrait;
 trait ControllerHelperTrait
 {
 	use ContainerAwareTrait;
+	use RedirectTrait {
+		redirect as redirectResponse;
+	}
 
 	/**
 	 * Returns a file response container.
 	 */
+	#[Deprecated('create and return a "mako\http\response\senders\File" instance instead', since: 'Mako 12.1.0')]
 	protected function fileResponse(string $file): File
 	{
 		return new File($this->fileSystem, $file);
 	}
 
 	/**
-	 * Returns a redirect response container.
-	 */
-	protected function redirectResponse(string $location, array $routeParams = [], array $queryParams = [], string $separator = '&', bool|string $language = true): Redirect
-	{
-		if ($this->routes->hasNamedRoute($location)) {
-			$location = $this->urlBuilder->toRoute($location, $routeParams, $queryParams, $separator, $language);
-		}
-
-		return new Redirect($location);
-	}
-
-	/**
 	 * Returns a stream response container.
 	 */
+	#[Deprecated('create and return a "mako\http\response\senders\Stream" instance instead', since: 'Mako 12.1.0')]
 	protected function streamResponse(Closure $stream, ?string $contentType = null, ?string $charset = null): Stream
 	{
 		return new Stream($stream, $contentType, $charset ?? $this->response->getCharset());
@@ -52,6 +45,7 @@ trait ControllerHelperTrait
 	/**
 	 * Returns a JSON response builder.
 	 */
+	#[Deprecated('create and return a "mako\http\response\builders\JSON" instance instead', since: 'Mako 12.1.0')]
 	protected function jsonResponse(mixed $data, int $options = 0, ?int $status = null, ?string $charset = null): JSON
 	{
 		return new JSON($data, $options, $status ?? $this->response->getStatus(), $charset ?? $this->response->getCharset());